pyproject.toml
[tool.poetry]
name = "high-order-networks-torch"
version = "0.1.0"
description = "Standard networks using high order layers"
authors = ["jloveric <john.loverich@gmail.com>"]
license = "MIT"

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
high-order-layers-torch = "^1.2.6"
torchvision = "^0.14.1"
omegaconf = "^2.3.0"
pytorch-lightning = "^1.8.6"

[tool.poetry.dev-dependencies]
hydra-core = "^1.3.1"
hydra-nevergrad-sweeper = "1.3.0.dev0"
black = "^22.12.0"


[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
